+++ src/main/java/com/nisovin/shopkeepers/playershops/PlayerInactivity.java
@@ -1,10 +1,11 @@
 package com.nisovin.shopkeepers.playershops;
 
 import java.util.ArrayList;
-import java.util.HashSet;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
-import java.util.UUID;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import org.bukkit.Bukkit;
@@ -13,10 +14,13 @@ import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitTask;
 
 import com.nisovin.shopkeepers.SKShopkeepersPlugin;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperRegistry;
+import com.nisovin.shopkeepers.api.events.PlayerInactiveEvent;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
+import com.nisovin.shopkeepers.api.user.User;
 import com.nisovin.shopkeepers.config.Settings;
+import com.nisovin.shopkeepers.shopkeeper.SKShopkeeperRegistry;
 import com.nisovin.shopkeepers.util.bukkit.SchedulerUtils;
+import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 import com.nisovin.shopkeepers.util.java.Validate;
 import com.nisovin.shopkeepers.util.logging.Log;
 
@@ -95,66 +99,99 @@ public class PlayerInactivity {
 
 		Log.info("Checking for shopkeepers of inactive players.");
 
-		Set<UUID> shopOwnerIds = new HashSet<>();
-		ShopkeeperRegistry shopkeeperRegistry = plugin.getShopkeeperRegistry();
+		// Get all shop owners:
+		// This initially contains all shop owners (and maps to no shopkeepers), and is then pruned.
+		Map<User, List<PlayerShopkeeper>> shopsByInactivePlayers = new HashMap<>();
+		SKShopkeeperRegistry shopkeeperRegistry = plugin.getShopkeeperRegistry();
 		shopkeeperRegistry.getAllPlayerShopkeepers().forEach(playerShop -> {
-			shopOwnerIds.add(playerShop.getOwnerUUID());
+			// In this first step, we are only interested in the existing shop owners, and therefore don't store the
+			// shopkeepers. We later gather the shopkeepers of only the inactive shop owners.
+			// We insert the empty list here instead of null, because the subsequent code needs to be able to
+			// differentiate between there being a mapping for a specific user versus the user being mapped to null.
+			shopsByInactivePlayers.put(playerShop.getOwnerUser(), Collections.emptyList());
 		});
-		if (shopOwnerIds.isEmpty()) {
+		if (shopsByInactivePlayers.isEmpty()) {
 			return; // There are no player shops
 		}
 
-		// Fetch OfflinePlayers async:
-		int playerShopkeeperInactiveDays = Settings.playerShopkeeperInactiveDays;
+		// Retrieve the OfflinePlayers and their 'last played' times asynchronously:
 		Bukkit.getScheduler().runTaskAsynchronously(plugin, () -> {
-			Set<UUID> inactiveShopOwnerIds = new HashSet<>();
-			long nowMillis = System.currentTimeMillis();
-			for (UUID shopOwnerId : shopOwnerIds) {
-				OfflinePlayer offlinePlayer = Bukkit.getOfflinePlayer(shopOwnerId);
-				// Some servers may delete player data files, either regularly for all players (which breaks this
-				// feature), or for particular players (for example to reset or fix some issue with their data). If this
-				// is the case, we cannot reliably determine when the player was last seen on the server, and therefore
-				// don't delete their player shopkeepers.
-				if (!offlinePlayer.hasPlayedBefore()) continue;
-
-				long lastPlayedMillis = offlinePlayer.getLastPlayed();
-				if (lastPlayedMillis == 0) continue; // 0 if unknown (see reasoning above)
-
-				long millisSinceLastPlayed = nowMillis - lastPlayedMillis;
-				int daysSinceLastPlayed = (int) TimeUnit.MILLISECONDS.toDays(millisSinceLastPlayed);
-				if (daysSinceLastPlayed >= playerShopkeeperInactiveDays) {
-					inactiveShopOwnerIds.add(shopOwnerId);
+			// Note: We don't use #removeIf(Predicate) here, because the implementation may evaluate the Predicate
+			// multiple times (which may be costly).
+			Iterator<User> shopOwnersIterator = shopsByInactivePlayers.keySet().iterator();
+			while (shopOwnersIterator.hasNext()) {
+				User shopOwner = shopOwnersIterator.next();
+				if (!isInactive(shopOwner)) {
+					shopOwnersIterator.remove();
 				}
 			}
-
-			if (inactiveShopOwnerIds.isEmpty()) {
-				// No inactive players found:
-				return;
+			if (shopsByInactivePlayers.isEmpty()) {
+				return; // No inactive players found
 			}
 
 			// Continue in main thread:
 			SchedulerUtils.runTaskOrOmit(plugin, () -> {
-				List<PlayerShopkeeper> shopsByInactivePlayers = new ArrayList<>();
+				// Get the shopkeepers owned by the inactive players:
 				shopkeeperRegistry.getAllPlayerShopkeepers().forEach(playerShop -> {
-					// If the shop is owned by an inactive player, mark it for removal:
-					UUID ownerId = playerShop.getOwnerUUID();
-					if (inactiveShopOwnerIds.contains(ownerId)) {
-						shopsByInactivePlayers.add(playerShop);
+					// If the shop is owned by an inactive player, remember it for removal:
+					User shopOwner = playerShop.getOwnerUser();
+					List<PlayerShopkeeper> ownedShopkeepers = shopsByInactivePlayers.computeIfPresent(shopOwner, (user, shopkeepers) -> {
+						// Replace the initial unmodifiable empty list with a modifiable one:
+						if (shopkeepers.isEmpty()) {
+							shopkeepers = new ArrayList<>();
+						}
+						return shopkeepers;
+					});
+					if (ownedShopkeepers != null) {
+						ownedShopkeepers.add(playerShop);
 					}
 				});
 
-				// Delete those shopkeepers:
+				// Process the inactive players and their shopkeepers:
 				if (!shopsByInactivePlayers.isEmpty()) {
-					shopsByInactivePlayers.forEach(playerShop -> {
-						Log.info("Deleting shopkeeper " + playerShop.getIdString() + " at " + playerShop.getPositionString()
-								+ " owned by " + playerShop.getOwnerString() + " due to owner inactivity.");
-						playerShop.delete();
+					shopsByInactivePlayers.entrySet().forEach(entry -> {
+						User user = entry.getKey();
+						List<PlayerShopkeeper> shopkeepers = entry.getValue();
+						int originalShopkeepersCount = shopkeepers.size();
+						assert !shopkeepers.isEmpty();
+
+						// Call event:
+						PlayerInactiveEvent event = new PlayerInactiveEvent(user, shopkeepers);
+						Bukkit.getPluginManager().callEvent(event);
+						if (event.isCancelled() || shopkeepers.isEmpty()) {
+							Log.debug("Ignoring inactive player " + TextUtils.getPlayerString(user) + " and their " + originalShopkeepersCount
+									+ (shopkeepers.size() == originalShopkeepersCount ? "" : " (reduced to " + shopkeepers.size() + ")")
+									+ " shopkeepers: Cancelled by a plugin.");
+						} else {
+							// Delete the shopkeepers:
+							shopkeepers.forEach(playerShop -> {
+								Log.info("Deleting shopkeeper " + playerShop.getIdString() + " at " + playerShop.getPositionString()
+										+ " owned by " + playerShop.getOwnerString() + " due to owner inactivity.");
+								playerShop.delete();
+							});
+						}
 					});
 
-					// Save:
-					plugin.getShopkeeperStorage().save();
+					// Save if necessary:
+					plugin.getShopkeeperStorage().saveIfDirty();
 				}
 			});
 		});
+	}
+
+	private boolean isInactive(User user) {
+		OfflinePlayer offlinePlayer = user.getOfflinePlayer();
+		// Some servers may delete player data files, either regularly for all players (which breaks this feature), or
+		// for particular players (for example to reset or fix some issue with their data). If this is the case, we
+		// cannot reliably determine when the player was last seen on the server, and therefore do not delete their
+		// shopkeepers.
+		if (!offlinePlayer.hasPlayedBefore()) return false;
+
+		long lastPlayedMillis = offlinePlayer.getLastPlayed();
+		if (lastPlayedMillis == 0) return false; // 0 if unknown (see reasoning above)
+
+		long millisSinceLastPlayed = System.currentTimeMillis() - lastPlayedMillis;
+		int daysSinceLastPlayed = (int) TimeUnit.MILLISECONDS.toDays(millisSinceLastPlayed);
+		return (daysSinceLastPlayed >= Settings.playerShopkeeperInactiveDays);
 	}
 }